packages/modeldb: rename api.ts to ModelAPI.ts, create RelationAPI.ts
