+++ packages/modeldb-pg/src/ModelAPI.ts
@@ -20,18 +20,9 @@ import {
 	validateModelValue,
 } from "@canvas-js/modeldb"
 
-import { PostgresPrimitiveValue, Encoder, Decoder } from "./encoding.js"
-import { Method, Query } from "./utils.js"
-
-const columnTypes = {
-	integer: "BIGINT",
-	float: "DOUBLE PRECISION",
-	number: "DOUBLE PRECISION",
-	string: "TEXT",
-	bytes: "BYTEA",
-	boolean: "BOOLEAN",
-	json: "JSONB",
-} satisfies Record<PrimitiveType, string>
+import { RelationAPI } from "./RelationAPI.js"
+import { PostgresPrimitiveValue, Encoder, Decoder, columnTypes } from "./encoding.js"
+import { Method, Query, quote } from "./utils.js"
 
 function getColumn(name: string, type: PrimitiveType, nullable: boolean) {
 	if (nullable) {
@@ -40,8 +31,6 @@ function getColumn(name: string, type: PrimitiveType, nullable: boolean) {
 		return `"${name}" ${columnTypes[type]} NOT NULL`
 	}
 }
-
-const quote = (name: string) => `"${name}"`
 
 export class ModelAPI {
 	public static async create(client: pg.Client, config: Config, model: Model, clear: boolean = false) {
@@ -711,150 +700,5 @@ export class ModelAPI {
 		} else {
 			return `${filters.map((filter) => `(${filter})`).join(" AND ")}`
 		}
-	}
-}
-
-export class RelationAPI {
-	public readonly table: string
-
-	readonly #select: Query
-	readonly #insert: Method
-	readonly #delete: Method
-	readonly #clear: Method
-
-	public static async create(client: pg.Client, config: Config, relation: Relation, clear: boolean) {
-		const sourceColumnNames: string[] = []
-		const targetColumnNames: string[] = []
-
-		const sourcePrimaryKey = config.primaryKeys[relation.source]
-		const targetPrimaryKey = config.primaryKeys[relation.target]
-
-		const columns: string[] = []
-
-		if (sourcePrimaryKey.length === 1) {
-			const [{ type }] = sourcePrimaryKey
-			columns.push(`_source ${columnTypes[type]} NOT NULL`)
-			sourceColumnNames.push("_source")
-		} else {
-			for (const [i, { type }] of sourcePrimaryKey.entries()) {
-				const columnName = `_source/${i}`
-				columns.push(`"${columnName}" ${columnTypes[type]} NOT NULL`)
-				sourceColumnNames.push(columnName)
-			}
-		}
-
-		if (targetPrimaryKey.length === 1) {
-			const [{ type }] = targetPrimaryKey
-			columns.push(`_target ${columnTypes[type]} NOT NULL`)
-			targetColumnNames.push("_target")
-		} else {
-			for (const [i, { type }] of targetPrimaryKey.entries()) {
-				const columnName = `_target/${i}`
-				columns.push(`"${columnName}" ${columnTypes[type]} NOT NULL`)
-				targetColumnNames.push(columnName)
-			}
-		}
-
-		// Initialize tables
-		const api = new RelationAPI(client, config, relation, sourceColumnNames, targetColumnNames)
-
-		const queries = []
-
-		if (clear) {
-			queries.push(`DROP TABLE IF EXISTS "${api.table}"`)
-		}
-
-		const sourceIndex = `${relation.source}/${relation.sourceProperty}/source`
-		const targetIndex = `${relation.source}/${relation.sourceProperty}/target`
-		const sourceColumns = sourceColumnNames.map(quote).join(", ")
-		const targetColumns = targetColumnNames.map(quote).join(", ")
-
-		queries.push(`CREATE TABLE IF NOT EXISTS "${api.table}" (${columns.join(", ")})`)
-		queries.push(`CREATE INDEX IF NOT EXISTS "${sourceIndex}" ON "${api.table}" (${sourceColumns})`)
-		if (relation.indexed) {
-			queries.push(`CREATE INDEX IF NOT EXISTS "${targetIndex}" ON "${api.table}" (${targetColumns})`)
-		}
-
-		await client.query(queries.join(";\n"))
-
-		return api
-	}
-
-	private constructor(
-		readonly client: pg.Client,
-		readonly config: Config,
-		readonly relation: Relation,
-		readonly sourceColumnNames: string[],
-		readonly targetColumnNames: string[],
-	) {
-		this.table = `${relation.source}/${relation.sourceProperty}`
-
-		// Prepare methods
-		const insertColumns = [...sourceColumnNames, ...targetColumnNames].map(quote).join(", ")
-		const insertParams = Array.from({ length: sourceColumnNames.length + targetColumnNames.length })
-			.map((_, i) => `$${i + 1}`)
-			.join(", ")
-		this.#insert = new Method(client, `INSERT INTO "${this.table}" (${insertColumns}) VALUES (${insertParams})`)
-
-		const selectBySource = sourceColumnNames.map((name, i) => `"${name}" = $${i + 1}`).join(" AND ")
-
-		this.#delete = new Method(client, `DELETE FROM "${this.table}" WHERE ${selectBySource}`)
-		this.#clear = new Method(client, `DELETE FROM "${this.table}"`)
-
-		// Prepare queries
-		const targetColumns = targetColumnNames.map(quote).join(", ")
-		this.#select = new Query(client, `SELECT ${targetColumns} FROM "${this.table}" WHERE ${selectBySource}`)
-
-		// this.#insert = `INSERT INTO "${this.table}" (_source, _target) VALUES ($1, $2)`
-		// this.#delete = `DELETE FROM "${this.table}" WHERE _source = $1`
-		// this.#select = `SELECT _source, _target FROM "${this.table}" WHERE _source = $1`
-		// this.#selectMany = `SELECT _source, _target FROM "${this.table}" WHERE _source = ANY($1)`
-	}
-
-	public async get(sourceKey: PostgresPrimitiveValue[]): Promise<PostgresPrimitiveValue[][]> {
-		const targets = await this.#select.all(sourceKey)
-		return targets.map((record) => this.targetColumnNames.map((name) => record[name]))
-	}
-
-	public async add(sourceKey: PostgresPrimitiveValue[], targetKeys: PostgresPrimitiveValue[][]) {
-		for (const targetKey of targetKeys) {
-			await this.#insert.run([...sourceKey, ...targetKey])
-		}
-	}
-
-	public async delete(sourceKey: PostgresPrimitiveValue[]) {
-		await this.#delete.run(sourceKey)
-	}
-
-	public async clear() {
-		await this.#clear.run([])
-	}
-
-	public async getMany(sources: PostgresPrimitiveValue[][]): Promise<PostgresPrimitiveValue[][][]> {
-		return await Promise.all(sources.map((source) => this.get(source)))
-
-		// // postgres doesn't know at planning time if the array has a single string
-		// let queryResult: pg.QueryResult<{ _source: PrimaryKeyValue; _target: PrimaryKeyValue }>
-		// if (sources.length === 0) {
-		// 	return []
-		// } else if (sources.length === 1) {
-		// 	queryResult = await this.client.query<{ _source: PrimaryKeyValue; _target: PrimaryKeyValue }, [PrimaryKeyValue]>(
-		// 		this.#select,
-		// 		[sources[0]],
-		// 	)
-		// } else {
-		// 	queryResult = await this.client.query<
-		// 		{ _source: PrimaryKeyValue; _target: PrimaryKeyValue },
-		// 		[PrimaryKeyValue[]]
-		// 	>(this.#selectMany, [sources])
-		// }
-
-		// const results: Record<string, string[]> = {}
-		// for (const row of queryResult.rows) {
-		// 	results[row._source] ||= []
-		// 	results[row._source].push(row._target)
-		// }
-
-		// return results
 	}
 }
