+++ packages/modeldb-sqlite/src/ModelDB.ts
@@ -17,7 +17,7 @@ import {
 	PrimaryKeyValue,
 } from "@canvas-js/modeldb"
 
-import { ModelAPI } from "./api.js"
+import { ModelAPI } from "./ModelAPI.js"
 
 export interface ModelDBOptions {
 	path: string | null
